+++ .eslintrc.js
@@ -1,0 +1,11 @@
+module.exports = {
+  extends: "ericbiewener-typescript",
+  env: {
+    browser: false,
+    es6: true,
+    node: true
+  },
+  globals: {
+    __non_webpack_require__: false
+  }
+};