+++ src/createCompletionItemProvider.ts
@@ -1,14 +1,6 @@
-// FIXME: notify user that they should turn off auto-import JS & TS if it is on now that Vandelay supports this
-import {
-  CancellationToken,
-  CompletionItem,
-  CompletionItemKind,
-  Position,
-  TextDocument,
-  TextEdit,
-} from 'vscode'
+import { CompletionItem, CompletionItemKind, Position, TextDocument, TextEdit } from 'vscode'
 import { cacheFileManager } from './cacheFileManager'
-import { MergedExportData, Plugin, PluginConfig, RichQuickPickItem } from './types'
+import { Plugin, PluginConfig, RichQuickPickItem } from './types'
 
 /**
  * Although we are not supposed to modify `CompletionItem.additionalTextEdits` inside
@@ -16,18 +8,20 @@ import { MergedExportData, Plugin, PluginConfig, RichQuickPickItem } from './typ
  * as a codebase grows. It's not slow... Mumbai took ~150-250 ms, but that's 3-5 times slower than
  * deferring that work until `resolveCompletionItems`.
  *
- * Therefore, we do the work in `resolveCompletionItems`. Although it is apparently possible that
- * the additional text edit will not complete by the time the CompletionItem is resolved, I am
+ * Therefore, we do the work in `resolveCompletionItems`. Although it's apparently possible that the
+ * additional text edit will not complete by the time the CompletionItem is resolved, I am
  * consistently seeing resolve times in *under 1ms*! Seems pretty much impossible to select a
  * CompletionItem before that resolution completes.
  *
  * If we eventually find that sometimes the resolution truly doesn't complete, we can add the
  * `selectImportForActiveWord` command as a backup: https://bit.ly/2Yn8xDA. This works well because
- * it is essentially a no-op if the additionalTextEdit insertion was successful.
+ * it is essentially a no-op if the additionalTextEdit insertion was successful. The downside is
+ * that it adds an additional action to the undo stack.
  *
  * Note that clicking a suggestion does NOT cause it to auto-import, but that's not a failure of
  * Vandelay's implementation -- VS Code simply doesn't trigger anything additional beyond the
- * insertion of the clicked word (i.e. even CompletionItem.command doesn't run).
+ * insertion of the clicked word (i.e. even CompletionItem.command doesn't run). This is the same as
+ * with VS Code's default autoImport functionality.
  */
 
 type RichCompletionItem<Q = RichQuickPickItem> = CompletionItem & {
@@ -42,7 +36,7 @@ export function createCompletionItemProvider(
   return {
     async provideCompletionItems(document: TextDocument, position: Position) {
       return await cacheFileManager(plugin, async exportData => {
-        if (!exportData) return [1]
+        if (!exportData) return []
 
         const mergedData = plugin.mergeExportData(exportData)
         const items = plugin.buildImportItems(plugin as any, mergedData, Object.keys(mergedData))
@@ -52,6 +46,7 @@ export function createCompletionItemProvider(
             item.label,
             CompletionItemKind.Event
           ) as RichCompletionItem
+          // Caching for use in `resolveCompletionItem`
           completionItem.importItem = item
           completionItem.position = position
           return completionItem