+++ src/plugins/javascript/cacheNodeModules.ts
@@ -5,16 +5,6 @@ import { cacheFileManager } from '../../cacheFileManager'
 import { isFile, writeCacheFile } from '../../utils'
 import { ExportDataJs, ExportDataNodeModulesJs, NodeModuleExports, PluginJs } from './types'
 
-// Watch for changes to package.json Go into each module listed there and look at its
-// package.json... grab the "main" property from that to get the entry file and just require() that
-// file to get all it's imports? This won't work for stuff that gets exported from deeper files but
-// not reexported from main file. Considering that, perhaps we end up doing a blended approach with
-// how things currently work.
-
-// And how does vs code handle this? For lodash-es, for example, is it able to get everything from
-// it's individual files? Perhaps there is a package.json standard for this, eg if you don't include
-// a "main" property then that means parse every top-level Js file?
-
 type Dep = Record<string, string>
 type PackageJson = { dependencies?: Dep; devDependencies?: Dep; peerDependencies?: Dep }
 
@@ -30,6 +20,7 @@ export async function cacheNodeModules(plugin: PluginJs) {
 
   const data: ExportDataNodeModulesJs = {}
   await Promise.all(jsons.map(j => cacheDependencies(plugin, j, data)))
+  console.log('DONE CACHING DEPENDENCY')
 
   await cacheFileManager(plugin, async cachedData => {
     const newData = cachedData as ExportDataJs
@@ -58,11 +49,14 @@ async function cacheDependencies(
   )
 }
 
-// FIXME: possible to make `plugin` global as well, like contet, so that i don't have to pass it around?
+// FIXME: possible to make `plugin` global as well, like context, so that i don't have to pass it
+// around? The difference would be that it gets set at the start of each command, whereas context
+// would never change.
 export async function cacheDependency(
   plugin: PluginJs,
   dep: string
 ): Promise<NodeModuleExports | undefined> {
+  console.log('CACHING DEPENDENCY')
   const dir = path.join(plugin.projectRoot, 'node_modules', dep)
   const packageJsonPath = path.join(dir, 'package.json')
   let packageJson
@@ -121,27 +115,25 @@ function getDefaultName(dep: string) {
 export function findPackageJsonFiles(plugin: PluginJs) {
   const results = []
   // First check if subdirectories of root have `package.json` files. If they do, we're looking at a
-  // workspace or monorepo of some kind and we probably want to parse the subdirectory modules
-  // rather than those in the project root.
+  // yarn workspace and we probably want to parse the subdirectory modules rather than those in the
+  // project root.
   for (const item of fs.readdirSync(plugin.projectRoot)) {
     if (item === 'node_modules' || item.startsWith('.')) continue
     const fullPath = path.join(plugin.projectRoot, item)
     if (isFile(fullPath)) continue
-    
-    const packageJsonPath = getPackageJsonPath(fullPath)
-    if (packageJsonPath) results.push(packageJsonPath)
-  }
 
-  // If no results yet, then look in project root.
-  if (!results.length) {
-    const packageJsonPath = getPackageJsonPath(plugin.projectRoot)
-    if (packageJsonPath) results.push(packageJsonPath)
+        t packageJsonPath = getPackageJsonPath(fullPath)
+        packageJsonPath) results.push(packageJsonPath)
   }
-
-  return results
+   // If no results yet, then look in project root.
+  if  !results.length) {
+        t packageJsonPath = getPackageJsonPath(plugin.projectRoot)
+        packageJsonPath) results.push(packageJsonPath)
+  }
+   re  rn results
 }
 
 function getPackageJsonPath(dir: string) {
-  const packageJsonPath = path.join(dir, 'package.json')
-  return isFile(packageJsonPath) ? packageJsonPath : null
+  co  t packageJsonPath = path.join(dir, 'package.json')
+  re  rn isFile(packageJsonPath) ? packageJsonPath : null
 }