+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ### Version 4.0.0
 - Import suggestions will now appear as you type!
-- `package.json` parsing for Improved caching of node modules
+- `package.json` parsing for Improved caching of node modules.
 - More fixes to command `Remove Unused Imports`.
+- Better removal of unused imports and importing of undefined variables in Typescript.
 
 **Breaking Changes**
 - Removed command `Import and Insert at Cursor`. It was just a lazy alternative to providing suggestions as you type.